backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
celery==5.2.2
colorama==0.4.4
django-cors-headers==3.10.0
django-celery-results==2.2.0
django-db-connection-pool==1.1.0
django-easy-audit==1.3.0
django-filter==21.1
django-redis==5.0.0
Django==3.1.13
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
# django-q==1.3.9
# drf-haystack==1.8.11
fake_useragent==0.1.11
# jieba==0.42.1
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
pyDes==2.0.1
pymysql==1.0.2
pytz==2021.3
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
wcwidth==0.2.5
# whoosh==2.7.4
PyJWT~=2.3.0
loguru==0.6.0
user-agents==2.2.0
uwsgi
drf-yasg==1.20.0
django-celery-beat==2.2.1
tenacity==8.0.1
ntplib==0.4.0
validators==0.18.2
django-simpleui==2022.4.9
pika==1.2.1
lxml==4.8.0
pymmh3==0.0.5
zoomeye==2.1.1
pika==1.2.1
PySocks==1.7.1
bs4==0.0.1
# selenium==4.1.5
mysqlclient==2.1.0